+++ packages/frontend/src/components/open/search.tsx
@@ -16,8 +16,7 @@ const Bar = styled.div`
   border: 1px solid ${colorGrayscale.gray400};
   border-radius: 40px;
   display: flex;
-  width: 520px;
-  height: 32px;
+  width: 560px;
   align-items: center;
   gap: 8px;
   background-color: ${colorOpacity['white_0.8']};
@@ -28,16 +27,18 @@ const Bar = styled.div`
     background-color: ${colorGrayscale.white};
   }
   ${mq.tabletOnly`
-    width: 440px;
+    width: 480px;
   `}
   ${mq.mobileOnly`
-    width: calc(100% - 48px);
+    width: 100%;
+    max-width: 480px;
   `}
 `
 const Input = styled.input`
   background-color: initial;
   flex: 1 0 0;
   line-height: 24px;
+  height: 24px;
   padding: 0 2px;
   color: ${colorGrayscale.gray800};
 
@@ -57,6 +58,7 @@ const SearchButton = styled(IconButton)`
   height: 24px;
 `
 
+const releaseBranch = process.env.NEXT_PUBLIC_RELEASE_BRANCH
 const SearchBar: React.FC = () => {
   return (
     <Bar>
@@ -64,7 +66,7 @@ const SearchBar: React.FC = () => {
       <SearchButton
         type={IconButton.Type.PRIMARY}
         theme={IconButton.THEME.normal}
-        iconComponent={<Search />} //todo: add release branch
+        iconComponent={<Search releaseBranch={releaseBranch} />}
       />
     </Bar>
   )