+++ packages/frontend/src/styles/z-index.ts
@@ -2,4 +2,5 @@ export enum ZIndex {
   Header = 1000,
   HamburgerMenu = 1100,
   Bar = Header,
+  FilterModal = 1200,
 }